refactor session code. tweak share button styling.
