+++ index.html
@@ -272,7 +272,8 @@
 
             <!-- Bookmark -->
             <li class="nav-item">
-                <button id="spacewalk-share-button" class="btn btn-outline-light btn-sm mt-1" type="button" data-toggle="modal" data-target="#spacewalk-share-url-modal">Share</button>
+                <!-- btn-outline-light btn-secondary -->
+                <button id="spacewalk-share-button" class="btn btn-secondary btn-sm ml-4 mt-1" type="button" data-toggle="modal" data-target="#spacewalk-share-url-modal">Share</button>
             </li>
 
         </ul>
